Copying Dev-VR to Dev-Main
#lockdown nick.penwarden

[CL 2791734 by Nick Whiting in Main branch]

+++ Engine/Source/Runtime/Core/Public/Misc/AutomationTest.h
@@ -325,7 +325,7 @@ protected:
 	}
 
 	/** For timers, track the first time this ticks */
-	float StartTime;
+	double StartTime;
 
 	friend class FAutomationTestFramework;
 };
